tmux: choosebuffer
